+++ source/browseMode.py
@@ -1076,6 +1076,18 @@ qn(
 	prevError=_("no previous error"),
 )
 qn(
+	"slider",
+	key="j",
+	# Translators: Input help message for a quick navigation command in browse mode.
+	nextDoc=_("moves to the next slider"),
+	# Translators: Message presented when the browse mode element is not found.
+	nextError=_("no next slider"),
+	# Translators: Input help message for a quick navigation command in browse mode.
+	prevDoc=_("moves to the previous slider"),
+	# Translators: Message presented when the browse mode element is not found.
+	prevError=_("no previous slider"),
+)
+qn(
 	"article",
 	key=None,
 	# Translators: Input help message for a quick navigation command in browse mode.