+++ user_docs/en/changes.md
@@ -8,6 +8,7 @@
 
 * A new command, assigned to `NVDA+x`, has been introduced to repeat the last information spoken by NVDA; pressing it twice shows it in a browseable message. (#625, @CyrilleB79)
 * Added an unassigned command to toggle keyboard layout. (#19211, @CyrilleB79)
+* Added a new Quick Navigation command `j` for jumping to next/previous slider in browse mode. (#17005, @hdzrvcc0X74)
 
 ### Changes